refactor(#30): Rename options.ts to trick-list-option and change for oriented object
