+++ public/views/options.html
@@ -76,7 +76,7 @@
         <h2>Catégories</h2>
     </section>
 
-    <script src="../options.js"></script>
+    <script src="../trick-list-options.js"></script>
 </body>
 
 </html>
